Split out add to cart logic to product-details to avoid errors.
